pyproject.toml
[project]
name = "clientapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.104.1",
  "uvicorn==0.23.2",
  "sqlalchemy==2.0.23",
  "psycopg2-binary==2.9.9",
  "python-dotenv==1.0.0",
  "pydantic==2.4.2",
]
